+++ src/main/scala/com/dataintuitive/viash/functionality/dataobjects/BooleanObject.scala
@@ -25,7 +25,7 @@ case class BooleanObjectRegular(
   name: String,
   alternatives: List[String] = Nil,
   description: Option[String] = None,
-  example: Option[String] = None,
+  example: Option[Boolean] = None,
   default: Option[Boolean] = None,
   required: Boolean = false,
   tag: Option[String] = None,
@@ -42,7 +42,7 @@ case class BooleanObjectRegular(
     name: String, 
     alternatives: List[String],
     description: Option[String],
-    example: Option[String],
+    example: Option[Boolean],
     default: Option[Boolean],
     required: Boolean,
     direction: Direction,
@@ -58,7 +58,6 @@ case class BooleanObjectTrue(
   name: String,
   alternatives: List[String] = Nil,
   description: Option[String] = None,
-  example: Option[String] = None,
   tag: Option[String] = None,
   direction: Direction = Input,
   oType: String = "boolean_true"
@@ -69,13 +68,13 @@ case class BooleanObjectTrue(
   val default: Option[Boolean] = None
   val multiple: Boolean = false
   val multiple_sep: Char = ':'
+  val example: Option[Boolean] = None
 
   def copyDO(
     oType: String, 
     name: String, 
     alternatives: List[String],
     description: Option[String],
-    example: Option[String],
     default: Option[Boolean],
     required: Boolean,
     direction: Direction,
@@ -91,7 +90,6 @@ case class BooleanObjectFalse(
   name: String,
   alternatives: List[String] = Nil,
   description: Option[String] = None,
-  example: Option[String] = None,
   tag: Option[String] = None,
   direction: Direction = Input,
   oType: String = "boolean_false"
@@ -102,13 +100,13 @@ case class BooleanObjectFalse(
   val default: Option[Boolean] = None
   val multiple: Boolean = false
   val multiple_sep: Char = ':'
+  val example: Option[Boolean] = None
 
   def copyDO(
     oType: String, 
     name: String, 
     alternatives: List[String],
     description: Option[String],
-    example: Option[String],
     default: Option[Boolean],
     required: Boolean,
     direction: Direction,