+++ src/main/scala/com/dataintuitive/viash/platforms/DebugPlatform.scala
@@ -42,6 +42,8 @@ case class DebugPlatform(
       arguments = functionality.arguments.map {
         case arg if arg.required && arg.default.isDefined => 
           arg.copyDO(required = false)
+        case arg if arg.required && arg.example.isDefined => 
+          arg.copyDO(required = false, default = arg.example)
         case arg: BooleanObject if arg.default.isEmpty => 
           arg.copyDO(required = false, default = Some(true))
         case arg: DoubleObject if arg.default.isEmpty =>