+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Viash 0.5.7
 
+## BREAKING CHANGES
+
+* `viash config`: An argument's example now needs to be of the same type as the argument itself. 
+  For example, `[ type: integer, name: foo, example: 10 ]` is valid, whereas 
+  `[ type: integer, name: foo, example: bar ]` is not, as 'bar' cannot be cast to an integer.
+
 ## NEW FUNCTIONALITY
 
 * `viash config inject`: A command for inserting a Viash header into your script.