+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 * `viash config inject`: A command for inserting a Viash header into your script.
 
 * `DockerPlatform`: Added a requirement setup for installing through yum. Example:
-  `setup: [ [ type: yum, packages: wget ] ]`
+  `setup: [ [ type: yum, packages: [ wget] ] ]`
+
+* `DockerPlatform`: Allow using copy and add instructions. Example:
+  `setup: [ [ type: docker, add: [ "http://foo.bar ." ]]]`
 
 ## BUG FIXES